Rename drive command to TeleopDrive, clean up constants and fix errors
