+++ src/main/java/frc/robot/subsystems/TalonDriveTrain.java
@@ -52,8 +52,8 @@ public class TalonDriveTrain extends SubsystemBase {
     public void setMotors(double left, double right){
         setLeftMotor(-left);
         setRightMotor(right);
-    }
-
+	}
+	
     public void setLeftMotor(double output){
 		leftMotor.set(ControlMode.PercentOutput, Math.max(-1, Math.min(1, output * speedMultiplier)));
 	}
